Reorganiza helpers do router de cards
